requirements.txt
python-telegram-bot==22.0 # for Telegram 
openai==1.51.0 # the OpenAI API
anthropic==0.31.2 # the Anthropic API
google-genai==1.5.0 # the Google GenAI API
Pillow==11.0.0 # for processing images